feat: refactoring upload screen
