+++ src/screens/upload/Upload.tsx
@@ -1,7 +1,16 @@
-import React, { useState } from 'react'
+import React, { useRef, useState } from 'react'
 
-import { useNavigation } from '@react-navigation/native'
-import { KeyboardAvoidingView, Platform, ScrollView, Text } from 'react-native'
+import ActionSheet from '@alessiocancian/react-native-actionsheet'
+import { RouteProp, useNavigation } from '@react-navigation/native'
+import {
+  ImageSourcePropType,
+  KeyboardAvoidingView,
+  Linking,
+  Platform,
+  ScrollView,
+  Text,
+  TouchableOpacity,
+} from 'react-native'
 import { useMutation } from 'react-query'
 import { useRecoilState, useRecoilValue } from 'recoil'
 
@@ -11,9 +20,17 @@ import { CloseButtonHeader } from '@/components/headers/closeButtonHeader/CloseB
 import { SelectArchivingModal } from '@/components/modal/selectArchivingModal/SelectArchivingModal'
 import { GrayTag } from '@/components/tag/grayTag/GrayTag'
 import i18n from '@/locales'
+import { ImageUploadMenuType, ImageUploadMenus } from '@/models/enums/ActionSheetType'
+import { Permissions } from '@/models/enums/Permissions'
 import { MainNavigationProp } from '@/navigations/MainNavigator'
+import { RootStackParamList } from '@/navigations/RootStack'
+import { handleContentImageMenu } from '@/services/ActionSheetService'
+import { createCancelConfirmAlert } from '@/services/Alert'
+import { checkAndRequestPermission } from '@/services/PermissionService'
+import { handleCameraOpen, handleFileOpen, handleImageSelect } from '@/services/imagePicker'
 import { SelectArchivingState } from '@/state/upload/SelectArchivingState'
 import { SelectTagState } from '@/state/upload/SelectTagState'
+import { colors } from '@/styles/colors'
 
 import {
   AddTagButton,
@@ -21,20 +38,27 @@ import {
   ArchivingSelect,
   Condition,
   Container,
+  Image,
+  PlusImageButton,
   RowView,
   Styles,
   TextInput,
   Title,
-} from '../Upload.style'
+} from './Upload.style'
+
+interface UploadProps {
+  route: RouteProp<RootStackParamList, 'Upload'>
+}
 
 /**
  *
  */
-export const LinkUpload = () => {
+export const Upload = ({ route }: UploadProps) => {
   const navigation = useNavigation<MainNavigationProp>()
   const [archivingName, setArchivingName] = useState('')
   const [contentName, setContentName] = useState('')
   const [link, setLink] = useState('')
+  const [image, setImage] = useState<ImageSourcePropType | ''>('')
   const [memo, setMemo] = useState('')
 
   const [openArchivingModal, setOpenArchivingModal] = useState(false)
@@ -46,12 +70,15 @@ export const LinkUpload = () => {
   const selectArchiving = useRecoilValue(SelectArchivingState)
   const [selectTag, setSelectTag] = useRecoilState(SelectTagState)
 
+  const actionSheetRef = useRef<ActionSheet>(null)
+
   const { mutate } = useMutation(() =>
     postContents({
-      contentType: 'link',
+      contentType: route.params.type,
       archivingId: 0,
       title: contentName,
       link: link,
+      imgUrl: '',
       tagIds: [],
       memo: memo,
     })
@@ -77,6 +104,13 @@ export const LinkUpload = () => {
    */
   const handleContentBlur = () => {
     setContentFocus(false)
+  }
+
+  /**
+   *
+   */
+  const handleUploadImage = () => {
+    actionSheetRef.current?.show()
   }
 
   /**
@@ -111,7 +145,19 @@ export const LinkUpload = () => {
    *
    */
   const handlesubmit = () => {
-    // TODO
+    // mutate()
+  }
+
+  /**
+   * handleActionSheetMenu
+   */
+  const handleActionSheetMenu = async (index: ImageUploadMenuType) => {
+    const selectImage = await handleContentImageMenu(index)
+
+    if (selectImage) {
+      console.log(selectImage)
+      setImage({ uri: selectImage })
+    }
   }
 
   return (
@@ -146,6 +192,8 @@ export const LinkUpload = () => {
       <Condition style={[contentName.length > 0 ? Styles.conditionComplete : null]}>
         {i18n.t('contentVerify')}
       </Condition>
+
+      {/* Link */}
       <Title>{i18n.t('link')}</Title>
       <TextInput
         placeholder={i18n.t('placeHolderLink')}
@@ -160,6 +208,26 @@ export const LinkUpload = () => {
       />
       {/* TODO: Condition Icon 추가 */}
       <Condition>{i18n.t('checkUrl')}</Condition>
+      {/* Image */}
+      <Title>{i18n.t('image')}</Title>
+      {image ? (
+        <TouchableOpacity onPress={handleUploadImage}>
+          <Image source={image} />
+        </TouchableOpacity>
+      ) : (
+        <PlusImageButton onPress={handleUploadImage}>
+          <Text>+</Text>
+        </PlusImageButton>
+      )}
+      <ActionSheet
+        ref={actionSheetRef}
+        title={i18n.t('uploadImage')}
+        options={ImageUploadMenus()}
+        cancelButtonIndex={0}
+        tintColor={colors.gray600}
+        onPress={handleActionSheetMenu}
+        theme="ios"
+      />
       <RowView>
         <Title>{i18n.t('tag')}</Title>
         <Text>{i18n.t('choice10')}</Text>
@@ -203,10 +271,11 @@ export const LinkUpload = () => {
           ]}
         />
       </KeyboardAvoidingView>
+      {/* TODO: 타입에 따른 분리 */}
       <BoxButton
         textKey={i18n.t('complete')}
         onPress={handlesubmit}
-        isDisabled={!archivingName || !contentName || !link}
+        isDisabled={!archivingName || !contentName || !image || !link}
       />
     </Container>
   )
